refactor & use validation package
