+++ modules/validation/helpers.go
@@ -10,6 +10,7 @@ import (
 	"regexp"
 	"strings"
 
+	user_model "code.gitea.io/gitea/models/user"
 	"code.gitea.io/gitea/modules/setting"
 )
 
@@ -112,4 +113,8 @@ func IsValidExternalTrackerURLFormat(uri string) bool {
 	}
 
 	return true
+}
+
+func IsValidUsername(user string) bool {
+	return user_model.IsUsableUsername(user) == nil
 }
